Move Filter definition stuff to Events.Store
