+++ Source/Events.Store/Streams/Filters/TypeFilterWithEventSourcePartitionDefinition.cs
@@ -3,13 +3,11 @@
 
 using System.Collections.Generic;
 using Dolittle.Artifacts;
-using Dolittle.Runtime.Events.Store;
-using Dolittle.Runtime.Events.Store.Streams;
 
-namespace Dolittle.Runtime.Events.Processing.Filters
+namespace Dolittle.Runtime.Events.Store.Streams.Filters
 {
     /// <summary>
-    /// Represents the definition for <see cref="TypeFilterWithEventSourcePartition"/>.
+    /// Represents the <see cref="IFilterDefinition" /> for type filter with event source partition.
     /// </summary>
     public class TypeFilterWithEventSourcePartitionDefinition : IPersistableFilterDefinition
     {
